+++ src/tools/binary_crates.rs
@@ -1,5 +1,5 @@
 use crate::cmd::{Binary, Command, Runnable};
-use crate::tools::{Tool, CARGO_INSTALL_UPDATE};
+use crate::tools::Tool;
 use crate::{Toolchain, Workspace};
 use failure::Error;
 use std::path::PathBuf;
@@ -58,11 +58,25 @@ impl Tool for BinaryCrate {
         Ok(())
     }
 
+    #[cfg(not(feature = "unstable"))]
     fn update(&self, workspace: &Workspace, _fast_install: bool) -> Result<(), Error> {
-        Command::new(workspace, &CARGO_INSTALL_UPDATE)
+        Command::new(workspace, &crate::tools::CARGO_INSTALL_UPDATE)
             .args(&[self.crate_name])
             .timeout(None)
             .run()?;
+        Ok(())
+    }
+
+    #[cfg(feature = "unstable")]
+    fn update(&self, workspace: &Workspace, fast_install: bool) -> Result<(), Error> {
+        let mut cmd = Command::new(workspace, &Toolchain::MAIN.cargo())
+            .args(&["-Zinstall-upgrade", "install", self.crate_name])
+            .env("__CARGO_TEST_CHANNEL_OVERRIDE_DO_NOT_USE_THIS", "nightly")
+            .timeout(None);
+        if fast_install {
+            cmd = cmd.args(&["--debug"]);
+        }
+        cmd.run()?;
         Ok(())
     }
 }