+++ src/tools/mod.rs
@@ -11,6 +11,7 @@ use std::path::PathBuf;
 
 pub(crate) static RUSTUP: Rustup = Rustup;
 
+#[cfg(not(feature = "unstable"))]
 pub(crate) static CARGO_INSTALL_UPDATE: BinaryCrate = BinaryCrate {
     crate_name: "cargo-update",
     binary: "cargo-install-update",
@@ -31,6 +32,7 @@ pub(crate) static GIT_CREDENTIAL_NULL: BinaryCrate = BinaryCrate {
 
 static INSTALLABLE_TOOLS: &[&dyn Tool] = &[
     &RUSTUP,
+    #[cfg(not(feature = "unstable"))]
     &CARGO_INSTALL_UPDATE,
     &RUSTUP_TOOLCHAIN_INSTALL_MASTER,
     &GIT_CREDENTIAL_NULL,